Updated based on pull review.
