+++ .gitignore
@@ -1,8 +1,14 @@
 # Ignore Gradle project-specific cache directory
 .gradle
 
+#Ignore the .idea folder.
+.idea
+
 # Ignore Gradle build output directory
 build
+
+# Ignore the run folder.
+run
 
 .classpath
 .project
