--- src/main/java/com/ocelotlovesbirds/birdhaus/datagen/ItemTagsGen.java
@@ -1,34 +1,0 @@
-package com.ocelotlovesbirds.birdhaus.datagen;
-
-import com.ocelotlovesbirds.birdhaus.Core;
-import com.ocelotlovesbirds.birdhaus.setup.Registration;
-
-import net.minecraft.data.DataGenerator;
-import net.minecraft.data.tags.BlockTagsProvider;
-import net.minecraft.data.tags.ItemTagsProvider;
-import net.minecraftforge.common.Tags;
-import net.minecraftforge.common.data.ExistingFileHelper;
-
-public class ItemTagsGen extends ItemTagsProvider {
-
-    public ItemTagsGen(DataGenerator generator, BlockTagsProvider blockTags, ExistingFileHelper helper) {
-        super(generator, blockTags, Core.MODID, helper);
-    }
-
-    /**
-     * Add tags to the registered item.
-     */
-    @Override
-    protected void addTags() {
-        tag(Tags.Items.CHESTS_WOODEN)
-                .add(Registration.BIRDHOUSE_ITEM.get());
-    }
-
-    /**
-     * @return Get the name of the tag file.
-     */
-    @Override
-    public String getName() {
-        return "birdhaustags";
-    }
-}
