--- src/main/java/com/ocelotlovesbirds/birdhaus/datagen/LootTablesGen.java
@@ -1,23 +1,0 @@
-package com.ocelotlovesbirds.birdhaus.datagen;
-
-import com.ocelotlovesbirds.birdhaus.setup.Registration;
-
-import net.minecraft.data.DataGenerator;
-
-public class LootTablesGen extends BaseLootTableProvider {
-
-
-    public LootTablesGen(DataGenerator dataGeneratorIn) {
-        super(dataGeneratorIn);
-    }
-
-    /**
-     * Create tables for the birdhouse, compatible with silk-touch tools.
-     */
-    @Override
-    protected void addTables() {
-
-        lootTables.put(Registration.BIRDHOUSE_BLOCK.get(), createSilkTouchTable("birdhouse",
-                Registration.BIRDHOUSE_BLOCK.get(), Registration.BIRDHOUSE_ITEM.get(), 1, 1));
-    }
-}
